Tilføjet resten af facademetoder og unittest af samme

+++ src/main/java/facades/IPersonFacade.java
@@ -14,8 +14,8 @@ import java.util.List;
  */
 public interface IPersonFacade {
   public Person addPerson(String fName, String lName, String phone);  
-  public Person deletePerson(int id);  
-  public Person getPerson(int id);  
+  public Person deletePerson(long id);  
+  public Person getPerson(long id);  
   public List<Person> getAllPersons();  
   public Person editPerson(Person p);  
 }
